requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.4
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-yasg==1.21.7
exceptiongroup==1.1.2
fake-useragent==1.2.1
h11==0.14.0
idna==3.4
inflection==0.5.1
lxml==4.9.3
oauthlib==3.2.2
outcome==1.2.0
packaging==23.1
Pillow==10.0.0
pycparser==2.21
PyJWT==2.8.0
PySocks==1.7.1
python-dotenv==1.0.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
text-unidecode==1.3
transliterate==1.10.2
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==2.0.4
webdriver-manager==4.0.0
wsproto==1.2.0
